refactor: split tag manager component to more details component
