+++ src/feature/tag-manager/types.ts
@@ -1,0 +1,8 @@
+export const TAG_TYPES = {
+  CHARACTER: 'character',
+  PLACE: 'place',
+  TIME: 'time',
+  DATE: 'date',
+} as const;
+
+export type TagType = (typeof TAG_TYPES)[keyof typeof TAG_TYPES];
